requirements.txt
falcon==1.2.0
gunicorn==19.7.1
marshmallow==3.0.0b3
pytest==3.2.1
python-decouple==3.1
requests==2.18.4
SQLAlchemy==1.1.14